+++ resources/ReleaseNotes.md
@@ -1,3 +1,5 @@
 # Software
 
+- Add compression flag to `mag.Science`
+- Add settings file to specify name of `timetable` properties for `mag.Science`
 - Move definition of time constants to shared utility file `mag.time.Constant`